rendervulkan: Move CVulkanDevice to header
